feat: 결과페이지 검색창 구현 #11
검색 기능 구현 완료

개선 사항: 추천 검색어 박스 정렬 등 디자인 개선
