+++ css/result_styles.css
@@ -15,6 +15,7 @@ body {
 }
 
 header {
+    position: relative;
     background-color: var(--dark-bg);
     color: var(--highlight);
     padding: 20px;
@@ -58,6 +59,63 @@ h1 {
 
 #search-button:hover {
     background-color: var(--gold);
+}
+
+/* 추천 검색어 박스 */
+#suggestions {
+    display: none; /* 기본적으로 숨김 */
+    position: absolute;
+    top: calc(100% + 5px); /* 검색창 바로 아래 위치 */
+    left: 50%;
+    transform: translateX(-50%);
+    width: 100%;
+    max-width: 500px; /* 검색창 최대 폭과 동일 */
+    background-color: var(--light-bg);
+    border: 1px solid var(--accent);
+    border-radius: 5px;
+    z-index: 2000;
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+    /* overflow: hidden; */
+}
+
+/* 추천 검색어 박스 내부 */
+.suggestion-box {
+    display: flex; /* 이미지와 텍스트를 가로로 정렬 */
+    align-items: center;
+    padding: 10px;
+    margin-bottom: 5px;
+    background-color: #ffffff;
+    border-bottom: 1px solid #ddd;
+    cursor: pointer;
+    transition: background-color 0.3s ease;
+}
+
+/* 마지막 추천 검색어 박스의 테두리 제거 */
+.suggestion-box:last-child {
+    border-bottom: none;
+}
+
+/* 앨범 커버 이미지 스타일 */
+.suggestion-box img.suggestion-img {
+    width: 20%; /* 박스 너비의 20% */
+    height: 100%;
+    border-radius: 5px;
+    margin-right: 10px;
+    object-fit: contain; /* 비율 유지 */
+}
+
+/* 추천 검색어 텍스트 */
+.suggestion-box span {
+    flex: 1; /* 나머지 공간을 차지 */
+    font-size: 14px; /* 텍스트 크기 */
+    color: var(--dark-bg);
+    white-space: nowrap; /* 텍스트 한 줄로 유지 */
+    overflow: hidden; /* 텍스트 오버플로우 숨김 */
+    text-overflow: ellipsis; /* 오버플로우된 텍스트에 말줄임표 표시 */
+}
+
+.suggestion-box:hover {
+    background-color: #f0f0f0; /* 박스 hover 효과 */
 }
 
 .result-grid {
@@ -162,7 +220,7 @@ body.popup-active::before {
     width: 100vw;
     height: 100vh;
     background-color: rgba(0, 0, 0, 0.5); /* 어두운 반투명 */
-    z-index: 1500; /* 팝업 뒤에 배경 */
+    z-index: 2000; /* 팝업 뒤에 배경 */
 }
 
 .popup {
