+++ index.html
@@ -47,7 +47,7 @@
         </div>
     </div>
     
-    <script type="module" src="js/main_search.js"></script>
+    <script type="module" src="js/search.js"></script>
     <script type="module" src="js/main.js"></script>
     <script type="module" src="js/main_api.js"></script>
     <script type="module" src="js/genre.js"></script>
