+++ result.html
@@ -11,6 +11,8 @@
         <h1>Playmingo</h1>
         <input type="text" id="search-bar" placeholder="검색어를 입력하세요..." autocomplete="off">
         <button id="search-button">검색</button>
+        <!-- 추천 검색어 표시 영역 -->
+        <div id="suggestions"></div>
     </header>
 
     <main>
